panic: Print the function name after stack corruption
On kernels with KASLR, printing the address of the function is not
helpful because the name of the function cannot be accurately obtained
from System.map during a post-mortem analysis as a result of the KASLR
offset not being logged. This makes it extremely difficult to debug
stack corruption panics.

Fix the issue by looking up the symbol name corresponding to the
problematic address in kallsyms if available.

Before: [    1.016861] Kernel panic - not syncing: stack-protector:
Kernel stack is corrupted in: ffffff8aa825f7e8
After: [    1.023759] Kernel panic - not syncing: stack-protector:
Kernel stack is corrupted in: wg_device_init+0x0/0x70

Signed-off-by: Danny Lin <danny@kdrag0n.dev>
Signed-off-by: Dakkshesh <dakkshesh5@gmail.com>
Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>
Signed-off-by: //. <174224977+skye-tachyon@users.noreply.github.com>